Ahora al momento de ejecutar el respaldo se le debe pasar como parametro el dia que se desea restaurar, si no se le pasa nigun parametro en toma por defecto el dia anterior

+++ restaura_pg.py
@@ -34,40 +34,61 @@ ipservidor, nombrebasedatos, usuariobasedatos,\
         rutarespaldo, nombrearchivo,\
         clave, puerto = fc.opcion_consultar('POSTGRESQL_RESPALDO')
 
-os.environ['PGPASSWORD'] = clave[1]
-
-
 def restaura(fecha_a_restaurar):
     '''
     string numero
-    Fecha a restaurar contiene el numero del dia a restaurar   
+    Fecha a restaurar contiene el numero del dia a restaurar
     '''
     nombre_archivo = nombrearchivo[1] + fecha_a_restaurar + '_.sql'
     ruta_y_archivo = os.path.join(rutarespaldo[1], nombre_archivo)
 
-    #Primero se procede a eliminar la base de datos del respaldo
-    eliminar_bd =   'dropdb  -h %s -p %s -U %s  %s' %\
-            (ipservidor[1], puerto[1], usuariobasedatos[1], nombrebasedatos[1])
+    #Verificamos que exista el archivo a restaurar
+    if os.path.isfile(ruta_y_archivo):
+        #Primero se procede a eliminar la base de datos del respaldo
+        eliminar_bd = 'dropdb  -h %s -p %s -U %s  %s' %\
+                (ipservidor[1], puerto[1], usuariobasedatos[1], nombrebasedatos[1])
+        #Se crea nuevamente la base de datos vacia
+        crear_bd = 'createdb -O %s -h %s -p %s -U %s %s' %\
+        (usuariobasedatos[1], ipservidor[1], puerto[1], usuariobasedatos[1], nombrebasedatos[1])
 
-    #Se crea nuevamente la base de datos vacia
-    crear_bd = 'createdb -O %s -h %s -p %s -U %s %s' %\
-            (usuariobasedatos[1], ipservidor[1], puerto[1], usuariobasedatos[1], nombrebasedatos[1])
+        #Se restaura el respaldo 
+        restaurar_bd = 'pg_restore -i -h %s -p %s -U %s -d %s -v "%s" ' %\
+        (ipservidor[1], puerto[1], usuariobasedatos[1], nombrebasedatos[1], ruta_y_archivo)
+        try:
+            print eliminar_bd
+            os.system(eliminar_bd)
+            print crear_bd
+            os.system(crear_bd)
+            print restaurar_bd
+            os.system(restaurar_bd)
+        except:
+            print "Ocurrio un error al momento de Restaurar la Base de datos"
+    else:
+        print 'No se encontro el archivo %s' % (ruta_y_archivo)
 
-    #Se restaura el respaldo 
-    restaurar_bd = 'pg_restore -i -h %s -p %s -U %s -d %s -v "%s" ' %\
-            (ipservidor[1], puerto[1], usuariobasedatos[1], nombrebasedatos[1], ruta_y_archivo)
+
+def main(opcion):
+    '''
+    Parametros recibidos 1: string numero del dia a restarurar.   
+    Si no se pasa ningun parametro el toma por defecto el respaldo
+    del dia anterior.
+    '''
     
-    try:
-        print eliminar_bd
-        os.system(eliminar_bd)
-        print crear_bd
-        os.system(crear_bd)
-        print restaurar_bd
-        os.system(restaurar_bd)
-    except:
-        print "Ocurrio un error al momento de Restaurar la Base de datos"
+    if len(opcion) <= 0:
+        fecha = str(datetime.date.today().day - 1) 
+    else:
+        fecha = opcion
+
+    print fecha
+    restaurar(fecha)
+
 
 if __name__ == '__main__':
-    #Day-1 Porque see restaura el respaldo del dia anterior
-    restaura(str(datetime.date.today().day-1))
+    try:
+        recibido = sys.argv[1]
+    except:
+        recibido = ''   
+    
+    os.environ['PGPASSWORD'] = clave[1]
+    main(recibido)
     os.system('unset PGPASSWORD')
